+++ app/styles/app.scss
@@ -4,6 +4,7 @@
 @import "components/status";
 @import "components/controls";
 @import "components/timer";
+@import "components/leaderboard";
 
 .app {
   margin-bottom: 20px;
