+++ app/styles/components/_leaderboard.scss
@@ -1,0 +1,7 @@
+.leaderboard {
+  border: 1px solid #545252;
+}
+
+.leaderboard td,th {
+  padding: 0.5rem;
+}
